Created my Django GymSession and Workout models
I think I correctly created the models for GymSession and Workout models. For the GymSession model, I created views that allows the user to get their scheduled gym sessions and the user should be able to schedule new gym sessions as long as there isnt an existing gym session for a specific day.