Completed LogWorkoutPage and HistoryPage
Completed LogWorkoutPage and HistoryPage. Users are now able to log their workout sessions for the current day. Users are also now able to checkout previous gym sessions, displaying the workouts during those sessions.

This project is almost completed. What is left is to include update and deletion capabilities as well as form validation, but before that, I am going to try and attempt to host this project online.