Fixing an issue when a user registers onto the page
When a user registers on the website, it automatically brings them to the home page. In HomePage.js, I have a forEach function that loops through the user friends, but it doesn't work if the user is new and doesn't have any friends which causes the home page to crash. To fix that, I added an if statement to check if the user friends exist.